Move mode and setMode higher
