+++ packages/components/src/components/SquiggleOutputViewer/index.tsx
@@ -1,7 +1,7 @@
-import { forwardRef, useState } from "react";
+import { forwardRef } from "react";
 
-import { SqOutputResult } from "../../../../squiggle-lang/src/public/types.js";
 import { SquiggleOutput } from "../../lib/hooks/useSquiggle.js";
+import { ViewerMode } from "../../lib/utility.js";
 import { CodeEditorHandle } from "../CodeEditor/index.js";
 import { PartialPlaygroundSettings } from "../PlaygroundSettings.js";
 import {
@@ -17,34 +17,17 @@ type Props = {
   squiggleOutput: SquiggleOutput;
   isRunning: boolean;
   editor?: CodeEditorHandle;
+  setMode: (mode: ViewerMode) => void;
+  mode: ViewerMode;
 } & PartialPlaygroundSettings;
-
-export type ViewerMode = "Imports" | "Exports" | "Variables" | "Result" | "AST";
-
-function useMode(outputResult: SqOutputResult) {
-  return useState<ViewerMode>(() => {
-    // Pick the initial mode value
-
-    if (!outputResult.ok) {
-      return "Variables";
-    }
-
-    const output = outputResult.value;
-    if (output.result.tag !== "Void") {
-      return "Result";
-    }
-    if (!output.exports.isEmpty()) {
-      return "Exports";
-    }
-    return "Variables";
-  });
-}
 
 /* Wrapper for SquiggleViewer that shows the rendering stats and isRunning state. */
 export const SquiggleOutputViewer = forwardRef<SquiggleViewerHandle, Props>(
-  ({ squiggleOutput, isRunning, editor, ...settings }, viewerRef) => {
+  (
+    { squiggleOutput, isRunning, editor, mode, setMode, ...settings },
+    viewerRef
+  ) => {
     const { output } = squiggleOutput;
-    const [mode, setMode] = useMode(output);
 
     return (
       <ViewerProvider
